requirements.txt
flask
flask-sqlalchemy
flask-login
markdown
python-dotenv
werkzeug
marked
